pyproject.toml
[project]
name = "pdf2md"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "jimingSu", email = "231880303@smail.nju.edu.cn" }
]
requires-python = ">=3.14"
dependencies = [
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]

[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"
